requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-bootstrap3==1.2.2
aldryn-categories==1.1.0
aldryn-common==1.0.4
-e git://github.com/aldryn/aldryn-newsblog.git@b5fc38e1a4dbda19539a8be9b8a603ca74b6c28f#egg=aldryn_newsblog
aldryn-people==2.1.0
aldryn-reversion==1.1.0
aldryn-translation-tools==0.2.1
Babel==2.5.1
backport-collections==0.1
certifi==2017.4.17
chardet==3.0.4
cmsplugin-filer==1.1.3
construct==2.5.3
dj-database-url==0.4.2
Django==1.10.8
django-appconf==1.0.2
django-appdata==0.1.6
django-ckeditor==5.3.0
django-classy-tags==0.8.0
django-cms==3.4.5
django-filebrowser-no-grappelli==3.7.2
django-filer==1.3.0
django-formtools==2.1
django-hvad==1.8.0
django-imagekit==4.0.1
django-js-asset==0.1.1
django-material==1.0.0
django-mptt==0.8.7
django-parler==1.9.1
django-phonenumber-field==1.3.0
django-polymorphic==1.3
django-reversion==1.10.2
django-revision==0.1.6
django-sekizai==0.10.0
django-sortedm2m==1.5.0
django-taggit==0.22.2
django-treebeard==4.2.0
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-installer==0.9.7
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.4.2
flickrapi==2.3
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.7
haystack==0.42
html5lib==0.9999999
idna==2.5
lxml==4.1.1
material==0.1
oauthlib==2.0.2
olefile==0.44
pefile==2017.9.3
phonenumbers==8.8.8
phonenumberslite==8.8.8
pilkit==2.0
Pillow==5.0.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-ptrace==0.9.2
python-slugify==1.2.0
pytz==2017.3
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.11.0
smmap2==2.0.3
solid-i18n==1.4.1
sqlite3client==0.0.1
tzlocal==1.5.1
Unidecode==0.4.21
urllib3==1.21.1
uWSGI==2.0.15
virtualenv==15.1.0
Whoosh==2.7.4
YURL==0.13
